requirements.txt
opencv-python==4.5.1.48
numpy==1.19.5
pygame==2.0.1
tensorflow==2.4.1
Pillow==8.2.0
Flask==1.1.2
Keras==2.4.3
matplotlib==3.4.1